separated file editor (monaco) to a component
